Changed Vector2Int.Random to return an int Added .Random to arrays and lists

+++ Core/Extensions/ExtensionMethods/VectorMathExtensions.cs
@@ -38,7 +38,7 @@ namespace ExtensionMethods
 			return UnityEngine.Random.Range(v2.x, maxInclusive: v2.y);
 		}
 
-		public static float Random(this Vector2Int v2)
+		public static int Random(this Vector2Int v2)
 		{
 			return UnityEngine.Random.Range(v2.x, v2.y);
 		}
